Hiking activity not finished, but push for nehemie to help me discover how to make the API link with the map asyncronous

+++ app/src/main/java/com/lastaoutdoor/lasta/data/model/Relation.kt
@@ -22,8 +22,8 @@ class Relation(
 }
 
 data class Bounds(
-    @SerializedName("minlat") @Expose val minlat: Float,
-    @SerializedName("minlon") @Expose val minlon: Float,
-    @SerializedName("maxlat") @Expose val maxlat: Float,
-    @SerializedName("maxlon") @Expose val maxlon: Float
+    @SerializedName("minlat") @Expose val minlat: Double,
+    @SerializedName("minlon") @Expose val minlon: Double,
+    @SerializedName("maxlat") @Expose val maxlat: Double,
+    @SerializedName("maxlon") @Expose val maxlon: Double
 )
